Properly recognize edge cases - in this particular implementation we are not throwing badspec errors because these type of CSS declarations seem to be up in the air as to how to interpret them.
Things like

*div
div*
div*div

are not clearly understood in the community. It's pretty easy to find web pages telling people that they can use these as either the grandchild or universal selector, and there is little consensus between the pages.

Firefox does not give an error message when parsing a selector built like the above, and instead it returns nothing - we are currently mimicking this functionality.

I almost feel like we are going to need a flag here to determine whether we are going to throw an error or not depending on bad selectors or not... some people might want it some way, but I definitely want it to continue processing for our own use.

Andy previously wrote it to throw exceptions though...