regarding last comment in last commit:
The multi * operator can be interpreted as either a universal "*" operator, or a grandparent operator.

The original version of this code attempted to trap relationship operators that started with * - which makes sense for the true relationship operators, but not for this particular operator since it has multiple possible uses.

If you remove this constraint it appears to behave correctly...