in all where we have not previously queried for a selector there was an issue whereby querying for speciicity would have caused query to do a lot of work for no reason.
For example you have a big tree loaded, you only want the specificity of a selector, not the elements that match the selector. In this case there was no way to avoid performing a heavy "query".

In these cases we do a quick substitute of the tree after which specificity uses a tiny tree to perform the calculation. Should save time in some cases